% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-variable-detection.R
\name{print.gbkmr_detection}
\alias{print.gbkmr_detection}
\title{Print method for gbkmr_detection objects}
\usage{
\method{print}{gbkmr_detection}(x, ...)
}
\arguments{
\item{x}{A gbkmr_detection object}

\item{...}{Additional arguments (not used)}
}
\value{
Invisible x
}
\description{
Print method for gbkmr_detection objects
}
