% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-user-interface.R
\name{print.gbkmr_results}
\alias{print.gbkmr_results}
\title{Print method for gbkmr_results objects}
\usage{
\method{print}{gbkmr_results}(x, ...)
}
\arguments{
\item{x}{A gbkmr_results object from \code{\link{gbkmr_run}}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisible x
}
\description{
Print method for gbkmr_results objects
}
