% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-user-interface.R
\name{summary.gbkmr_results}
\alias{summary.gbkmr_results}
\title{Summary method for gbkmr_results objects}
\usage{
\method{summary}{gbkmr_results}(object, ...)
}
\arguments{
\item{object}{A gbkmr_results object from \code{\link{gbkmr_run}}}

\item{...}{Additional arguments (not used)}
}
\value{
Invisible object
}
\description{
Summary method for gbkmr_results objects
}
